requirements.txt
nltk==3.8.1
pyspellchecker==0.7.2
pandas==2.0.3
pandarallel==1.6.5
numpy==1.23.5
xgboost==1.7.6
